Tests new 'smart_window_surroundings' options
This adds a new test case for the new values of the
'smart_window_surroundings' setting introduced in #1588.

Also, I've migrated of the existing test cases to new attribute
getters/setters. By using the new geometry attributes of clients, I got
rid of the x11 dependency in one of the existing test cases.