CI: remove old 32 bit job (#1614)
It does not seem to be possible to let github actions pass data into old docker containers. So let us disable the old 32 bit build for the time being.
Also, this commit updates the version of the github action modules and installs imagemagick and librsvg2-bin in the website job to make the CI pass again.