Make mergify compatible with up-to-date branch protection (#1641)
This change attempts to make mergify compatible with our branch
protection setting setting that pull requests are 'up to date' before
merging. I want this setting because it adds the 'Update branch' button
in pull requests that are not ahead of the current master branch.

I hope the present change to the mergify config achieves this because
mergify showed this error message:

> The branch protection setting 'Require branches to be up to date
> before merging' is not compatible with draft PR checks. To keep this
> branch protection enabled, update your Mergify configuration to enable
> in-place checks: set merge_queue.max_parallel_checks: 1, set every queue
> rule batch_size: 1, and avoid two-step CI (make merge_conditions
> identical to queue_conditions). Otherwise, disable this branch
> protection.