Add if_multiple_empty option to show_frame_decorations (#1565)
This adds an `if_multiple_empty` choice for `show_frame_decorations`. It works like `if_empty` to show a placeholder for childless frames, but it hides that placeholder on empty tags that only have a root frame. A similar behavior could only be achieved before with `show_frame_decorations` set to `if_empty` and `smart_frame_surroundings` on.